Antes de usar bootstrap

+++ app/Http/Controllers/EmpleadoController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Empleado;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage; 
 
 class EmpleadoController extends Controller
 {
@@ -41,7 +42,8 @@ class EmpleadoController extends Controller
             $datosEmpleado['Foto']=$request->file('Foto')->store('uploads','public');
         }
         Empleado::insert($datosEmpleado);
-        return response()->json($datosEmpleado);
+        //return response()->json($datosEmpleado);
+        return redirect('empleado')->with('mensaje','Empleado agregado con éxito');
     }
 
     /**
@@ -77,10 +79,16 @@ class EmpleadoController extends Controller
     public function update(Request $request, $id)
     {
         $datosEmpleado = request()->except(['_token','_method']);
+        if($request->hasFile('Foto')){
+            $empleado=Empleado::findOrFail($id);
+            Storage::delete('public/'.$empleado->Foto);
+            $datosEmpleado['Foto']=$request->file('Foto')->store('uploads','public');
+        }
+
+
         Empleado::where('id','=',$id)->update($datosEmpleado);
-        
         $empleado=Empleado::findOrFail($id);
-        return view('empleado.edit', compact('empleado'));
+        return view('empleado.edit', compact('empleado'));       
     }
 
     /**
@@ -91,7 +99,12 @@ class EmpleadoController extends Controller
      */
     public function destroy($id)
     {
-        Empleado::destroy($id);
-        return redirect('empleado');
+        $empleado=Empleado::findOrFail($id);
+
+        if( Storage::delete('public/'.$empleado->Foto)){
+            Empleado::destroy($id);
+        }        
+        //return redirect('empleado');
+        return redirect('empleado')->with('mensaje','Empleado eliminado');
     }
 }
